Move @main to executable target to fix Linux build
