+++ Sources/PGExtras/Commands/Bloat.swift
@@ -4,14 +4,16 @@ import ArgumentParser
 import TextTable
 
 
-struct Bloat: AsyncParsableCommand {
+public struct Bloat: AsyncParsableCommand {
     @OptionGroup var options: PGExtras.Options
 
-    func run() async throws {
+    public func run() async throws {
         try await Self.run(Row.Values.self,
                            credentials: options.credentials,
                            Row.init)
     }
+
+    public init() { }
 }
