+++ Sources/PGExtras/Commands/CacheHit.swift
@@ -4,14 +4,16 @@ import ArgumentParser
 import TextTable
 
 
-struct CacheHit: AsyncParsableCommand {
-    @OptionGroup var options: PGExtras.Options
+public struct CacheHit: AsyncParsableCommand {
+    @OptionGroup var options: Options
 
-    func run() async throws {
+    public func run() async throws {
         try await Self.run(Row.Values.self,
                            credentials: options.credentials,
                            Row.init)
     }
+
+    public init() { }
 }
 
 extension CacheHit: PGExtrasCommand {
