+++ Sources/PGExtras/Commands/IndexSize.swift
@@ -2,14 +2,16 @@ import ArgumentParser
 import TextTable
 
 
-struct IndexSize: AsyncParsableCommand {
+public struct IndexSize: AsyncParsableCommand {
     @OptionGroup var options: PGExtras.Options
 
-    func run() async throws {
+    public func run() async throws {
         try await Self.run(Row.Values.self,
                            credentials: options.credentials,
                            Row.init)
     }
+
+    public init() { }
 }
