+++ Sources/PGExtras/Commands/LongRunningQueries.swift
@@ -3,14 +3,16 @@ import TextTable
 import Foundation
 
 
-struct LongRunningQueries: AsyncParsableCommand {
+public struct LongRunningQueries: AsyncParsableCommand {
     @OptionGroup var options: PGExtras.Options
 
-    func run() async throws {
+    public func run() async throws {
         try await Self.run(Row.Values.self,
                            credentials: options.credentials,
                            Row.init)
     }
+
+    public init() { }
 }
