+++ Sources/PGExtras/Commands/RecordsRank.swift
@@ -2,14 +2,16 @@ import ArgumentParser
 import TextTable
 
 
-struct RecordsRank: AsyncParsableCommand {
+public struct RecordsRank: AsyncParsableCommand {
     @OptionGroup var options: PGExtras.Options
 
-    func run() async throws {
+    public func run() async throws {
         try await Self.run(Row.Values.self,
                            credentials: options.credentials,
                            Row.init)
     }
+
+    public init() { }
 }
