+++ Sources/PGExtras/Commands/SeqScans.swift
@@ -2,14 +2,16 @@ import ArgumentParser
 import TextTable
 
 
-struct SeqScans: AsyncParsableCommand {
+public struct SeqScans: AsyncParsableCommand {
     @OptionGroup var options: PGExtras.Options
 
-    func run() async throws {
+    public func run() async throws {
         try await Self.run(Row.Values.self,
                            credentials: options.credentials,
                            Row.init)
     }
+
+    public init() { }
 }
