+++ Sources/PGExtras/Commands/StatReset.swift
@@ -2,13 +2,15 @@ import ArgumentParser
 import TextTable
 
 
-struct StatReset: AsyncParsableCommand {
+public struct StatReset: AsyncParsableCommand {
     @OptionGroup var options: PGExtras.Options
 
-    func run() async throws {
+    public func run() async throws {
         try await Self.runQuery(credentials: options.credentials)
         print("pg_stat has been reset.")
     }
+
+    public init() { }
 }
