+++ Sources/PGExtras/Options.swift
@@ -1,0 +1,7 @@
+import ArgumentParser
+
+
+struct Options: ParsableArguments {
+    @Option(name: .shortAndLong, help: "Path to Postgres DB credentials.")
+    var credentials: Credentials
+}
